refactor: extract peer discovery into shared module
Address PR #2 review feedback:

1. Extract waitForNewPeer + getKnownDeviceIds into src/peer-discovery.ts,
   used by both init and invite. Eliminates duplicated peer polling logic
   and the less-robust version in init that didn't check known devices.

2. Type the spinner parameter as Ora instead of any.

Net: -34 lines, single source of truth for peer acceptance.
